[rest] Move the REST API code inside prosoul app
